feat: ChatRoom, ChatRoomParticipant
