+++ main/src/docs/asciidoc/index.adoc
@@ -22,7 +22,7 @@ include::group.adoc[]
 
 include::groupApplication.adoc[]
 
-include::participant.adoc[]
+include::groupParticipant.adoc[]
 
 include::jwt.adoc[]
