+++ main/src/docs/asciidoc/participant.adoc
@@ -26,18 +26,18 @@ operation::participation-history[snippets='query-parameters,http-request,http-re
 
 === 참여자 모각코 온도 올리기
 
-operation::participant-increase-temperature[snippets='path-parameters,http-request,http-response']
+operation::groupParticipant-increase-temperature[snippets='path-parameters,http-request,http-response']
 
 === 참여자 모각코 온도 낮추기
 
-operation::participant-decrease-temperature[snippets='path-parameters,http-request,http-response']
+operation::groupParticipant-decrease-temperature[snippets='path-parameters,http-request,http-response']
 
 === 참여자 탈퇴
 
-operation::participant-withdraw[snippets='path-parameters,http-request,http-response']
+operation::groupParticipant-withdraw[snippets='path-parameters,http-request,http-response']
 
 === 참여자 지각 요청
 
-operation::participant-tardiness[snippets='path-parameters,http-request,http-response']
+operation::groupParticipant-tardiness[snippets='path-parameters,http-request,http-response']
