+++ main/src/main/java/com/amorgakco/backend/chatroom/controller/ChatRoomController.java
@@ -1,0 +1,35 @@
+package com.amorgakco.backend.chatroom.controller;
+
+import com.amorgakco.backend.chatroom.domain.ChatRoom;
+import com.amorgakco.backend.chatroom.dto.ChatRoomPageResponse;
+import com.amorgakco.backend.chatroom.service.ChatRoomService;
+import com.amorgakco.backend.global.argumentresolver.AuthMember;
+import com.amorgakco.backend.global.argumentresolver.AuthMemberId;
+import com.amorgakco.backend.member.domain.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/chat-room")
+@RequiredArgsConstructor
+public class ChatRoomController {
+
+    private final ChatRoomService chatRoomService;
+
+    @GetMapping
+    public ChatRoomPageResponse getChatRoomList(@AuthMemberId final Member member,
+                                                @RequestParam final Integer page){
+        return chatRoomService.getChatRoomList(member,page);
+    }
+
+    @PostMapping("/{chatRoomId}")
+    public void enterChatRoom(@AuthMember final Member member, @PathVariable final Long chatRoomId){
+        chatRoomService.
+    }
+}
