+++ main/src/main/java/com/amorgakco/backend/chatroom/domain/ChatRoom.java
@@ -1,0 +1,28 @@
+package com.amorgakco.backend.chatroom.domain;
+
+
+import com.amorgakco.backend.global.BaseTime;
+import com.amorgakco.backend.group.domain.Group;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class ChatRoom extends BaseTime {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne
+    private Group group;
+
+
+}
