+++ main/src/main/java/com/amorgakco/backend/chatroom/dto/ChatRoomResponse.java
@@ -1,0 +1,9 @@
+package com.amorgakco.backend.chatroom.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ChatRoomResponse(
+        Long chatRoomId,
+        String groupName) {
+}
