+++ main/src/main/java/com/amorgakco/backend/chatroom/repository/ChatRoomRepository.java
@@ -1,0 +1,9 @@
+package com.amorgakco.backend.chatroom.repository;
+
+import com.amorgakco.backend.chatroom.domain.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ChatRoomRepository extends JpaRepository<ChatRoom,Long> {
+}
