+++ main/src/main/java/com/amorgakco/backend/global/exception/ErrorCode.java
@@ -44,7 +44,10 @@ public enum ErrorCode {
     INVALID_DIAGONAL_DISTANCE("313", "잘못된 그룹 검색 요청입니다. 위도 경도를 확인해주세요."),
     NOT_SAME_GROUP_PARTICIPANT("314", "동일한 그룹 내의 참여자끼리 온도를 변경할 수 있습니다."),
     EXCEED_PARTICIPATION_LIMIT("315", "활동중인 모각코는 5개를 넘을 수 없습니다."),
-    GROUP_APPLICATION_DUPLICATED("316", "중복된 참여 요청입니다");
+    GROUP_APPLICATION_DUPLICATED("316", "중복된 참여 요청입니다"),
+
+    //Chat Room Error Code
+    CHAT_ROOM_NOT_FOUND("400","채팅방이 존재하지 않습니다.");
 
     private final String code;
     private final String message;
