+++ main/src/main/java/com/amorgakco/backend/groupapplication/service/GroupApplicationValidator.java
@@ -5,7 +5,7 @@ import com.amorgakco.backend.global.exception.ParticipantException;
 import com.amorgakco.backend.group.domain.Group;
 import com.amorgakco.backend.groupapplication.repository.GroupApplicationRepository;
 import com.amorgakco.backend.member.domain.Member;
-import com.amorgakco.backend.participant.repository.ParticipantRepository;
+import com.amorgakco.backend.groupparticipant.repository.GroupParticipantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -16,11 +16,11 @@ import java.time.LocalDateTime;
 public class GroupApplicationValidator {
 
     private static final Integer PARTICIPATION_LIMIT = 5;
-    private final ParticipantRepository participantRepository;
+    private final GroupParticipantRepository groupParticipantRepository;
     private final GroupApplicationRepository groupApplicationRepository;
 
     public void validate(final Group group, final Member member) {
-        Integer participationCount = participantRepository.countCurrentParticipationByMember(member, LocalDateTime.now());
+        Integer participationCount = groupParticipantRepository.countCurrentParticipationByMember(member, LocalDateTime.now());
         validateParticipationLimit(participationCount);
         validateDuplicatedApplication(group, member);
     }
