+++ main/src/main/java/com/amorgakco/backend/groupparticipant/domain/LocationVerificationStatus.java
@@ -1,0 +1,6 @@
+package com.amorgakco.backend.groupparticipant.domain;
+
+public enum LocationVerificationStatus {
+    UNVERIFIED,
+    VERIFIED
+}
