+++ main/src/main/java/com/amorgakco/backend/groupparticipant/dto/GroupParticipationHistoryResponse.java
@@ -1,0 +1,14 @@
+package com.amorgakco.backend.groupparticipant.dto;
+
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record GroupParticipationHistoryResponse(
+        int page,
+        int elementSize,
+        boolean hasNext,
+        List<GroupParticipationHistory> histories
+) {
+}
