+++ main/src/main/java/com/amorgakco/backend/groupparticipant/dto/TemperatureResponse.java
@@ -1,0 +1,4 @@
+package com.amorgakco.backend.groupparticipant.dto;
+
+public record TemperatureResponse(Integer temperature) {
+}
