--- main/src/main/java/com/amorgakco/backend/participant/domain/LocationVerificationStatus.java
@@ -1,6 +1,0 @@
-package com.amorgakco.backend.participant.domain;
-
-public enum LocationVerificationStatus {
-    UNVERIFIED,
-    VERIFIED
-}
