--- main/src/main/java/com/amorgakco/backend/participant/dto/TemperatureResponse.java
@@ -1,4 +1,0 @@
-package com.amorgakco.backend.participant.dto;
-
-public record TemperatureResponse(Integer temperature) {
-}
