+++ main/src/main/resources/static/docs/index.html
@@ -1010,7 +1010,7 @@ Vary: Access-Control-Request-Headers</code></pre>
 <h4 id="_참여자_위치_인증_http_request"><a class="link" href="#_참여자_위치_인증_http_request">HTTP request</a></h4>
 <div class="listingblock">
 <div class="content">
-<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">PATCH /participants/locations HTTP/1.1
+<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">PATCH /groupParticipants/locations HTTP/1.1
 Content-Type: application/json;charset=UTF-8
 Content-Length: 75
 Host: localhost:8080
@@ -1071,7 +1071,7 @@ Vary: Access-Control-Request-Headers</code></pre>
 <h4 id="_참여자의_모각코_참여_내역_http_request"><a class="link" href="#_참여자의_모각코_참여_내역_http_request">HTTP request</a></h4>
 <div class="listingblock">
 <div class="content">
-<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /participants/histories?page=0 HTTP/1.1
+<pre class="highlightjs highlight nowrap"><code class="language-http hljs" data-lang="http">GET /groupParticipants/histories?page=0 HTTP/1.1
 Host: localhost:8080</code></pre>
 </div>
 </div>
@@ -1089,7 +1089,7 @@ Content-Length: 1489
 
 {
   "status" : "success",
-  "path" : "/participants/histories",
+  "path" : "/groupParticipants/histories",
   "data" : {
     "page" : 0,
     "elementSize" : 6,
