+++ main/src/test/java/com/amorgakco/backend/fixture/group/TestParticipantsFactory.java
@@ -1,11 +1,11 @@
 package com.amorgakco.backend.fixture.group;
 
+import com.amorgakco.backend.groupparticipant.domain.GroupParticipant;
 import com.amorgakco.backend.member.domain.Member;
-import com.amorgakco.backend.participant.domain.Participant;
 
 public class TestParticipantsFactory {
 
-    public static Participant create(final Member participant) {
-        return new Participant(participant);
+    public static GroupParticipant create(final Member participant) {
+        return new GroupParticipant(participant);
     }
 }
