+++ main/src/test/java/com/amorgakco/backend/fixture/participant/TestParticipantFactory.java
@@ -1,8 +1,8 @@
 package com.amorgakco.backend.fixture.participant;
 
 import com.amorgakco.backend.group.dto.LocationVerificationRequest;
-import com.amorgakco.backend.participant.dto.ParticipationHistory;
-import com.amorgakco.backend.participant.dto.ParticipationHistoryPagingResponse;
+import com.amorgakco.backend.groupparticipant.dto.GroupParticipationHistory;
+import com.amorgakco.backend.groupparticipant.dto.GroupParticipationHistoryResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,9 +16,9 @@ public class TestParticipantFactory {
     private static final LocalDateTime BEGIN_AT = LocalDateTime.now();
     private static final LocalDateTime END_AT = LocalDateTime.now().plusHours(3);
 
-    public static ParticipationHistoryPagingResponse currentParticipationHistoryResponse() {
-        List<ParticipationHistory> currentGroups = currentGroups();
-        return ParticipationHistoryPagingResponse.builder()
+    public static GroupParticipationHistoryResponse currentParticipationHistoryResponse() {
+        List<GroupParticipationHistory> currentGroups = currentGroups();
+        return GroupParticipationHistoryResponse.builder()
                 .page(0)
                 .hasNext(false)
                 .elementSize(currentGroups.size())
@@ -26,9 +26,9 @@ public class TestParticipantFactory {
                 .build();
     }
 
-    public static ParticipationHistoryPagingResponse pastParticipationHistoryResponse() {
-        List<ParticipationHistory> pastGroups = pastGroups();
-        return ParticipationHistoryPagingResponse.builder()
+    public static GroupParticipationHistoryResponse pastParticipationHistoryResponse() {
+        List<GroupParticipationHistory> pastGroups = pastGroups();
+        return GroupParticipationHistoryResponse.builder()
                 .page(0)
                 .hasNext(false)
                 .elementSize(pastGroups.size())
@@ -36,18 +36,18 @@ public class TestParticipantFactory {
                 .build();
     }
 
-    private static List<ParticipationHistory> currentGroups() {
+    private static List<GroupParticipationHistory> currentGroups() {
         return List.of(
                 participationHistory(1L), participationHistory(2L), participationHistory(3L));
     }
 
-    private static List<ParticipationHistory> pastGroups() {
+    private static List<GroupParticipationHistory> pastGroups() {
         return List.of(
                 participationHistory(4L), participationHistory(5L), participationHistory(6L));
     }
 
-    private static ParticipationHistory participationHistory(final Long groupId) {
-        return ParticipationHistory.builder()
+    private static GroupParticipationHistory participationHistory(final Long groupId) {
+        return GroupParticipationHistory.builder()
                 .groupId(groupId)
                 .address(ADDRESS)
                 .name(NAME)
